+++ CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## TODO
 
-- update localization for amnesia buffs (instruct player to seek treatment from traders, maybe even show treatment cost)
 - check out how forgetting elixir does its popup
+- update localization for amnesia buffs (instruct player to seek treatment from traders, maybe even show treatment cost)
 - update/replace all text referring to memory boosters
 
 ## [2.0.0]